+++ pepper/src/editor_utils.rs
@@ -459,7 +459,6 @@ pub static REGISTER_AUTO_MACRO: RegisterKey = RegisterKey::from_char_unchecked('
 pub static REGISTER_SEARCH: RegisterKey = RegisterKey::from_char_unchecked('s');
 pub static REGISTER_READLINE_PROMPT: RegisterKey = RegisterKey::from_char_unchecked('p');
 pub static REGISTER_READLINE_INPUT: RegisterKey = RegisterKey::from_char_unchecked('i');
-pub static REGISTER_COMMENT: RegisterKey = RegisterKey::from_char_unchecked('c');
 
 #[derive(Clone, Copy, PartialEq, Eq)]
 pub struct RegisterKey(u8);