+++ pepper/rc/changelog.md
@@ -7,6 +7,7 @@
 - changed search string will *also* be compiled into a pattern if it contains `^` or `$`
 - changed readline input and prompt are now registers `i` and `p` respectively
 - removed `@readline-input()` as the input can now be accessed by `register(i)`
+- added `toggle-comment` which will toggle a prefix comment for each line reached by a cursor
 
 # 0.29.0
 - added diagnostic logging to `spawn` command